requirements.txt
numpy==2.0.0
numpy-stl==3.1.1
pygame==2.6.0
python-utils==3.8.2
typing_extensions==4.12.2
